requirements.txt
pip~=23.0.1
attrs~=22.1.0
wheel~=0.37.1
django-request~=1.6.2
Jinja2~=3.1.2
setuptools~=60.2.0
packaging~=23.0
jsons~=1.6.3
Flask~=2.2.3
PyYAML~=6.0
psycopg2~=2.9.5
pandas~=2.0.0
scikit-learn~=1.2.2